Fix dedup tail issue.
